Complete type refactor.
Breaking changes:
Scale -> CalScale
ICalVersion -> Version
